refactor summation and array position

+++ A_Summation.c
@@ -9,11 +9,11 @@ int main()
     {
         scanf("%d",&arr[i]);
     }
-    int sum =0;
+    long long int sum =0;
     for (int i = 0; i < n; i++)
     {
         sum+=arr[i];
     }
-    printf("%d",sum < 0 ? -sum : sum);
+    printf("%lld",sum < 0 ? -sum : sum);
     return 0;
 }
